refactor: move remaining entities module
